move integration tests to /tests dir
