+++ src/main.rs
@@ -99,47 +99,3 @@ async fn main() -> anyhow::Result<()> {
 
     Ok(())
 }
-
-#[cfg(test)]
-mod test {
-    use assert_cmd::prelude::*;
-    use predicates::prelude::*;
-    use std::process::Command;
-
-    #[test]
-    fn help() -> Result<(), Box<dyn std::error::Error>> {
-        let mut cmd = Command::cargo_bin("rman")?;
-
-        cmd.arg("--help");
-        cmd.assert()
-            .success()
-            .stdout(predicate::str::contains("Usage: rman"));
-
-        Ok(())
-    }
-
-    #[test]
-    fn simple() -> Result<(), Box<dyn std::error::Error>> {
-        let mut cmd = Command::cargo_bin("rman")?;
-
-        cmd.arg("test/fixtures/ls_procfile");
-        cmd.assert()
-            .success()
-            .stderr(predicate::str::contains("ls -a"))
-            .stderr(predicate::str::contains("echo Hello World"));
-
-        Ok(())
-    }
-
-    #[test]
-    fn file_doesnt_exist() -> Result<(), Box<dyn std::error::Error>> {
-        let mut cmd = Command::cargo_bin("rman")?;
-
-        cmd.arg("test/file/doesnt/exist");
-        cmd.assert()
-            .failure()
-            .stderr(predicate::str::contains("No such file or directory"));
-
-        Ok(())
-    }
-}
