+++ tests/integration.rs
@@ -1,0 +1,40 @@
+use assert_cmd::prelude::*;
+use predicates::prelude::*;
+use std::process::Command;
+
+#[test]
+fn help() -> Result<(), Box<dyn std::error::Error>> {
+    let mut cmd = Command::cargo_bin("rman")?;
+
+    cmd.arg("--help");
+    cmd.assert()
+        .success()
+        .stdout(predicate::str::contains("Usage: rman"));
+
+    Ok(())
+}
+
+#[test]
+fn simple() -> Result<(), Box<dyn std::error::Error>> {
+    let mut cmd = Command::cargo_bin("rman")?;
+
+    cmd.arg("tests/fixtures/ls_procfile");
+    cmd.assert()
+        .success()
+        .stderr(predicate::str::contains("ls -a"))
+        .stderr(predicate::str::contains("echo Hello World"));
+
+    Ok(())
+}
+
+#[test]
+fn file_doesnt_exist() -> Result<(), Box<dyn std::error::Error>> {
+    let mut cmd = Command::cargo_bin("rman")?;
+
+    cmd.arg("tests/file/doesnt/exist");
+    cmd.assert()
+        .failure()
+        .stderr(predicate::str::contains("No such file or directory"));
+
+    Ok(())
+}
